refactor: FavoriteButton component
