+++ apps/Cybar/components/ReviewModal/OpenPosition/OpenPositionData.tsx
@@ -13,50 +13,25 @@ import {
   useOrderType,
   useTradeTpSl,
 } from "@symmio/frontend-sdk/state/trade/hooks";
-import {
-  useFavorites,
-  useLeverage,
-  useToggleUserFavoriteCallback,
-} from "@symmio/frontend-sdk/state/user/hooks";
+import { useLeverage } from "@symmio/frontend-sdk/state/user/hooks";
 import { OrderType } from "@symmio/frontend-sdk/types/trade";
 import { formatAmount, toBN } from "@symmio/frontend-sdk/utils/numbers";
 import { useGetTokenWithFallbackChainId } from "@symmio/frontend-sdk/utils/token";
 import Column from "components/Column";
-import { Star } from "components/Icons";
 import InfoItem from "components/InfoItem";
 import { DisplayLabel } from "components/InputLabel";
 import React, { useMemo } from "react";
 import styled, { useTheme } from "styled-components";
 import ActionButton from "./ActionButton";
+import { FavoriteButton } from "./FavoriteButton";
 
 const LabelsWrapper = styled(Column)`
   gap: 12px;
 `;
 
-const FavoriteButton = styled.button`
-  display: flex;
-  align-items: center;
-  padding: 5px 0;
-  width: 100%;
-  color: ${({ theme }) => theme.text3};
-  background-color: ${({ theme }) => theme.bg6};
-  border: ${({ theme }) => `1px solid ${theme.border2}`};
-  border-radius: 2px;
-  &:hover {
-    color: ${({ theme }) => theme.text2};
-    background-color: ${({ theme }) => theme.bg7};
-  }
-`;
-
-const FavoriteSpan = styled.span`
-  text-align: center;
-  width: 100%;
-`;
-
 export default function OpenPositionData() {
   const theme = useTheme();
   const { chainId } = useActiveWagmi();
-  const favorites = useFavorites();
 
   const orderType = useOrderType();
   const market = useActiveMarket();
@@ -67,11 +42,6 @@ export default function OpenPositionData() {
     chainId,
   );
   const { price, formattedAmounts } = useTradePage();
-
-  const isFavorite = market ? favorites?.includes(market.id) : false;
-  const toggleFavorite = market
-    ? useToggleUserFavoriteCallback(market.id)
-    : undefined;
 
   const [symbol, pricePrecision] = useMemo(
     () =>
@@ -179,21 +149,7 @@ export default function OpenPositionData() {
           key={info.title}
         />
       ))}
-      {toggleFavorite && (
-        <FavoriteButton onClick={toggleFavorite}>
-          <Star
-            size={16}
-            isFavorite={isFavorite}
-            style={{
-              marginLeft: "8px",
-            }}
-          />
-          <FavoriteSpan>
-            {" "}
-            {isFavorite ? "Remove from favorites" : "Add to favorites"}
-          </FavoriteSpan>
-        </FavoriteButton>
-      )}
+      {market && <FavoriteButton marketId={market.id} />}
       <ActionButton />
     </React.Fragment>
   );
